move engine client code to module
