+++ x/engine/client/auth.go
@@ -1,4 +1,4 @@
-package engine
+package client
 
 import (
 	"encoding/hex"
