+++ x/engine/client/auth_test.go
@@ -1,4 +1,4 @@
-package engine
+package client
 
 import (
 	"fmt"
