+++ x/engine/client/engine_client.go
@@ -1,4 +1,4 @@
-package engine
+package client
 
 import "context"
