+++ x/engine/client/types_test.go
@@ -1,4 +1,4 @@
-package engine
+package client
 
 import (
 	"encoding/json"
